Uber-Supply-Demand-Gap-Project
Uber Supply Demand Gap Project - The aim of analysis was to identify the root cause of the problem (i.e. cancellation and non-availability of cars) and recommend ways to improve the situation. This data set used was a masked data set which is similar to what data analysts at Uber handle.